+++ implants/Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["imix", "golem", "lib/eldritch", "lib/c2"]
+members = ["imix", "golem", "lib/eldritch", "lib/c2", "lib/windows_api"]
 resolver = "2"
 
 [workspace.dependencies]
@@ -70,6 +70,7 @@ tonic = { git = "https://github.com/hyperium/tonic.git", rev = "07e4ee1" }
 uuid = "1.5.0"
 which = "4.4.2"
 whoami = "1.3.0"
+windows_api = { path = "./lib/windows_api" }
 windows-sys = "0.45.0"
 winreg = "0.51.0"
