+++ implants/lib/windows_api/src/virtual_alloc_ex.rs
@@ -1,0 +1,40 @@
+use anyhow::Result;
+use {
+    object::LittleEndian as LE,
+    object::{Object, ObjectSection},
+    std::{os::raw::c_void, ptr::null_mut},
+    windows_sys::Win32::Security::SECURITY_ATTRIBUTES,
+    windows_sys::Win32::System::Threading::CreateRemoteThread,
+    windows_sys::Win32::{
+        Foundation::{GetLastError, BOOL, FALSE, HANDLE},
+        System::{
+            Diagnostics::Debug::WriteProcessMemory,
+            Memory::{
+                VirtualAllocEx, MEM_COMMIT, MEM_RESERVE, PAGE_EXECUTE_READWRITE,
+                PAGE_PROTECTION_FLAGS, VIRTUAL_ALLOCATION_TYPE,
+            },
+            Threading::{OpenProcess, PROCESS_ACCESS_RIGHTS, PROCESS_ALL_ACCESS},
+        },
+    },
+};
+// pub unsafe fn VirtualAllocEx(hprocess: super::super::Foundation::HANDLE, lpaddress: *const ::core::ffi::c_void, dwsize: usize, flallocationtype: VIRTUAL_ALLOCATION_TYPE, flprotect: PAGE_PROTECTION_FLAGS) -> *mut ::core::ffi::c_void
+fn virtual_alloc_ex(
+    hprocess: HANDLE,
+    lpaddress: *const c_void,
+    dwsize: usize,
+    flallocationtype: VIRTUAL_ALLOCATION_TYPE,
+    flprotect: PAGE_PROTECTION_FLAGS,
+) -> anyhow::Result<*mut c_void> {
+    let buffer_handle: *mut c_void =
+        unsafe { VirtualAllocEx(hprocess, lpaddress, dwsize, flallocationtype, flprotect) };
+    if buffer_handle == null_mut() {
+        let error_code = unsafe { GetLastError() };
+        if error_code != 0 {
+            return Err(anyhow::anyhow!(
+                "Failed to allocate memory. Last error returned: {}",
+                error_code
+            ));
+        }
+    }
+    Ok(buffer_handle)
+}
